requirements.txt
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.7.0.72
pandas==2.0.3
requests==2.31.0
Pillow==9.5.0
ultralytics==8.0.196
numpy==1.24.3